fix: CSharp was ignoring `--ignore-usings` option
PMDCPD configuration expects that a language class has some properties
set when passing options. We were passing the options to the configuration
object via `.setIgnoreUsings(true)` but this is not enough because the
`CSLanguage` class reads this option from a `java.util.Properties` object
which needs to be populated again with the correct property for ignore-using
to be used.
Moveover, when the language was not explicitly passed, or when the tool
was run multiple times with different languages, the analysis was performed
on all the files with all the supported languages.
So a CSharp file was first analyzed with the CSharp analyzer that
(after the properties fix) doesn't report issues when the `--ignore-usings`
flag is enabled, then the same CSharp file was re-analysed as a Java file,
and since the Java analyzer doesn't recognize the using syntax, the tool
was reporting the `using` code as a clone.
What we do now is to split the files by language and run every file
with just its specific language analyzer.